requirement.txt
fastapi
uvicorn
pydantic
passlib
bcrypt
python-jose
python-multipart
